+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Add ``Configuration.from_ini()`` method to load configuration from ini file.
+
 3.17.1
 ------
 - Fix ``DynamicContainer`` deep-copying bug.